isa_gus: improved 16-bit PCM (no whatsnew)
git-svn-id: svn://messdev.no-ip.org/mess@14410 9849079f-6229-0410-a07f-84406ce305ba

+++ src/mess/machine/isa_gus.c
@@ -266,8 +266,8 @@ void gf1_device::sound_stream_update(sound_stream &stream, stream_sample_t **inp
 						set_irq(IRQ_WAVETABLE,x);
 				}
 				if(m_voice[x].voice_ctrl & 0x20)
-				{  // 16-bit PCM (needs some work)
-					m_voice[x].sample = (INT16)((m_wave_ram[(current & 0xffffe)+1]) | ((m_wave_ram[current & 0xffffe])<<8)>>8);
+				{  // 16-bit PCM
+					m_voice[x].sample = (INT8)((m_wave_ram[(current & 0xffffe)+1]) | ((m_wave_ram[current & 0xffffe])<<8)>>8);
 				}
 				else
 				{  // 8-bit PCM
